+++ pyoffers/models/__init__.py
@@ -1,5 +1,6 @@
 # coding: utf-8
 from .advertiser import Advertiser, AdvertiserManager
+from .conversion import Conversion, ConversionManager
 from .core import ModelManager
 from .goal import Goal, GoalManager
 from .offer import Offer, OfferManager