+++ CHANGELOG.rst
@@ -3,6 +3,7 @@
 
 * Added support for returning multiple objects. (`#7`_)
 * Refactored models instantiation. (`#9`_)
+* Added `Conversion` model. (`#3`_)
 
 0.1.2
 =====
@@ -23,3 +24,4 @@
 .. _#9: https://github.com/Stranger6667/pyoffers/issues/9
 .. _#7: https://github.com/Stranger6667/pyoffers/issues/7
 .. _#5: https://github.com/Stranger6667/pyoffers/issues/5
+.. _#3: https://github.com/Stranger6667/pyoffers/issues/3